JDO Repository implementation added, JDO name as uppercase word rename trick

+++ greatage-domain/src/main/java/org/greatage/domain/AbstractEntityRepository.java
@@ -35,7 +35,7 @@ public abstract class AbstractEntityRepository implements EntityRepository {
 	}
 
 	public <PK extends Serializable, E extends Entity<PK>>
-	int count(final Class<E> entityClass) {
+	long count(final Class<E> entityClass) {
 		return count(entityClass, new AllCriteria<PK, E>());
 	}
